# go/model_link.go
/*
 * Shortlink Service
 *
 * It is an URL shortener service - it can generate a short version of arbitrary user's URL.
 * Also user can get statistics for URL generated url (how much people followed the link).
 *
 * API version: 1.0.0
 * Contact: ptsypyshev@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Link struct {
	ID int64 `json:"id,omitempty"`

	ShorkLink string `json:"shorkLink,omitempty"`

	LongLink string `json:"longLink,omitempty"`

	ClickCounter int64 `json:"clickCounter,omitempty"`

	Clickers []Clicker `json:"clickers,omitempty"`

	Status bool `json:"status,omitempty"`
}